updated folder structure of emoji
