+++ packages/nodes/emoji/src/handlers/index.ts
@@ -2,4 +2,7 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from './getEmojiOnInsert';
+export * from './getEmojiOnSelectItem';
+export * from './getFindTriggeringInput';
 export * from './getOnKeyDownEmoji';
