+++ packages/nodes/emoji/src/utils/EmojiPicker/EmojiPickerState.types.ts
@@ -1,0 +1,16 @@
+import { EmojiCategoryList } from '../../types';
+import { Emoji } from '../IndexSearch';
+
+export type EmojiPickerStateProps = {
+  searchValue: string;
+  hasFound: boolean;
+  isSearching: boolean;
+  searchResult: Emoji[];
+  emoji?: Emoji;
+  category: EmojiCategoryList;
+};
+
+export type EmojiPickerStateDispatch = {
+  type: string;
+  payload?: Partial<EmojiPickerStateProps>;
+};
