+++ packages/nodes/emoji/src/utils/EmojiPicker/index.ts
@@ -1,0 +1,8 @@
+/**
+ * @file Automatically generated by barrelsby.
+ */
+
+export * from './useEmojiPicker.types';
+export * from './useEmojiPicker';
+export * from './EmojiPickerState';
+export * from './EmojiPickerState.types';
