+++ packages/nodes/emoji/src/utils/IndexSearch/EmojiLibrary/EmojiInlineLibrary.types.ts
@@ -1,0 +1,1 @@
+export type THash = Record<string, string>;
