+++ packages/nodes/emoji/src/utils/IndexSearch/EmojiLibrary/index.ts
@@ -2,6 +2,8 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './EmojiInlineLibrary';
-export * from './EmojiFlyoutLibrary';
 export * from './EmojiLibrary.types';
+export * from './EmojiInlineLibrary';
+export * from './EmojiInlineLibrary.types';
+export * from './EmojiFlyoutLibrary.types';
+export * from './EmojiFlyoutLibrary';
